[#64] refactor: chat stats response record 정리
