+++ live-chat-server/src/main/java/org/giglab/live/application/dto/stats/ChatStatsResponse.java
@@ -1,3 +1,3 @@
-package org.giglab.live.application.dto.room;
+package org.giglab.live.application.dto.stats;
 
 public record ChatStatsResponse(int totalMessages, int totalQuestions, int aiAnswerCount) {}
